add condition if file exists

+++ application/controllers/admin/AdminAboutValuesController.php
@@ -208,8 +208,10 @@ class AdminAboutValuesController extends CI_Controller {
             if (!$response['updated']) {
                 $this->session->set_flashdata('error', $this->alert->show('Cannot update benefit', 0));
             } else {
-                if ($response['old_image_filename']) {
-                    unlink('images/values/'.$response['old_image_filename']);
+                if ($response['old_image_filename'] && !empty($_POST['image'])) {
+                    if (file_exists('images/values/'.$response['old_image_filename'])) {
+                        unlink('images/values/'.$response['old_image_filename']);
+                    }
                 }
                 $this->session->set_flashdata('success', $this->alert->show('Update success', 1));
             }
@@ -249,7 +251,9 @@ class AdminAboutValuesController extends CI_Controller {
             $this->session->set_flashdata('error', $this->alert->show('Cannot delete benefit', 0));
         } else {
             if ($response['old_image_filename']) {
-                unlink('images/values/'.$response['old_image_filename']);
+                if (file_exists('images/values/'.$response['old_image_filename'])) {
+                    unlink('images/values/'.$response['old_image_filename']);
+                }
             }
             $this->session->set_flashdata('success', $this->alert->show('Succecss delete!', 1));
         }
